requirements.txt
asgiref==3.8.1
certifi==2022.6.15
cffi==1.16.0
charset-normalizer==2.1.0
distutils644==0.4
Django==5.0.6
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
EasyProcess==1.1
entrypoint2==1.1
idna==3.3
mss==9.0.1
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
pybase64==1.3.2
pycparser==2.22
pygame==2.5.2
PyJWT==2.9.0
pynput==1.7.6
pyscreenshot==3.1
requests==2.28.1
setuptools==10.0
six==1.16.0
sounddevice==0.4.6
sqlparse==0.5.0
tzdata==2024.1
urllib3==1.26.11
